x & y should be uint16_t, not int..

+++ include/eastwood/FntFile.h
@@ -16,7 +16,7 @@ class FntFile
         ~FntFile();
 
         void extents(std::string text, uint16_t& w, uint16_t& h);
-        void render(std::string text, Surface &surface, int x, int y, uint8_t paloff);
+        void render(std::string text, Surface &surface, uint16_t x, uint16_t y, uint8_t paloff);
 
 
     private:
